requirements.txt
# Основные зависимости для управления квадропедом
# Python 3.7+ рекомендуется

# Управление сервоприводами
adafruit-circuitpython-servokit>=1.3.8
adafruit-blinka>=8.19.0  # Требуется для работы с CircuitPython библиотеками на Raspberry Pi

# Математические вычисления и кривые Безье
numpy>=1.19.0,<2.0.0
bezier>=2021.2.12

# Конфигурация
PyYAML>=5.4.0

# Управление с клавиатуры
keyboard>=0.13.5

# Работа с изображениями и видео
opencv-python>=4.5.0,<5.0.0
opencv-contrib-python>=4.5.0,<5.0.0  # Для дополнительных функций OpenCV
Pillow>=8.0.0

# Сетевая передача изображений
imagezmq>=1.1.1
pyzmq>=22.0.0

# Работа с видео потоками
imutils>=0.5.4

# Визуализация (для object_tracker_controller и модели)
matplotlib>=3.3.0

# Зависимости для object_tracker_controller (опционально)
# Требуют дополнительной настройки yolov4-deepsort-base проекта
tensorflow>=2.4.0,<2.15.0  # Для object_tracker_controller
absl-py>=0.12.0  # Для TensorFlow флагов

# Примечание: deep_sort и yolov4 требуют отдельной установки из репозитория:
# https://github.com/theAIGuysCode/yolov4-deepsort-base
# Эти модули не доступны через pip и должны быть установлены вручную

